Rename test_utils to utilities - caused confusion as it wasn't testing labkey.utils but rather provided "utilities" for the test infrastructure
